Crash in WebCore::RenderBlock::LineBreaker::nextLineBreak https://bugs.webkit.org/show_bug.cgi?id=93806
Patch by Douglas Stockwell <dstockwell@chromium.org> on 2012-08-13
Reviewed by Abhishek Arya.

Source/WebCore:

When looking for line breaks on the first line, existing code was
checking for text-combine only in the first-line style. Since
text-combine isn't inherited this resulted in a line break being
chosen before combineText was called. When this happened and then
combineText was called subsequently, the position of the line break
iterator would be invalid.

This patch changes the check to use the regular style as in
skipLeadingWhitespace and textWidth.

Test: fast/text/text-combine-first-line-crash.html

* rendering/RenderBlockLineLayout.cpp:
(WebCore::RenderBlock::LineBreaker::nextLineBreak): Don't use the
first-line style when checking text-combine.

LayoutTests:

* fast/text/text-combine-first-line-crash-expected.txt: Added.
* fast/text/text-combine-first-line-crash.html: Added.

Identifier: 111771@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125503 268f45cc-cd09-0410-ab3c-d52691b4dbfc