Speech Input: wrong position was reported for scolled-down elements https://bugs.webkit.org/show_bug.cgi?id=93634
Reviewed by Kent Tamura.

Source/WebCore:

When clicking on a speech-enabled input element, WebKit reports
the position of the element to the embedder, so that it can show
a bubble indicating that speech recognition is in progress for
the element that the user clicked on.

On a page that was scrolled, this position was being reported
erroneously: it was given relative to the frame content as opposed to
the root view, and the in-progress bubble would be shown in the wrong
place.

Test: fast/speech/bubble-position-scrolled.html

* html/shadow/TextControlInnerElements.cpp:
(WebCore::InputFieldSpeechButtonElement::startSpeechInput):

LayoutTests:

Test that the position of the speech element is passed to the embedder
corectly, i.e. where the user clicked, when the page is scrolled down.

* fast/speech/bubble-position-scrolled-expected.txt: Added.
* fast/speech/bubble-position-scrolled.html: Added.


Identifier: 111797@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125529 268f45cc-cd09-0410-ab3c-d52691b4dbfc