Delete some dead code in RenderBox::computePercentageLogicalHeight https://bugs.webkit.org/show_bug.cgi?id=94262
Reviewed by Tony Chang.

I'm pretty sure this is dead code. None of the layout tests hit this code,
including the ones added in the original patch http://trac.webkit.org/changeset/10983.
I spent a good while and could not concoct a test that hits this code.

The test-cases added in that file, as well as the test-cases I came up with that should
go through this codepath, instead go through computePositionedLogicalHeight and never
hit computePercentageLogicalHeight.
* rendering/RenderBox.cpp:
(WebCore::RenderBox::computePercentageLogicalHeight):

Identifier: 112189@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125938 268f45cc-cd09-0410-ab3c-d52691b4dbfc