Fix vsyncTick drought when vsync throttling is disabled https://bugs.webkit.org/show_bug.cgi?id=94012
Patch by Brian Anderson <brianderson@chromium.org> on 2012-08-17
Reviewed by James Robinson.

CCFrameRateController made an assumption that every vsyncTick results
in a call to didBeginFrame, which is not necessarily true and causes
the CCFrameRateController to stop ticking when vsync is disabled. We
move the manual ticks out of didBeginFrame and manually tick in the
proper place.

CCFrameRateControllerTest updated with use cases that should tick
without a didBeginFrame and will fail without this patch.

* platform/graphics/chromium/cc/CCFrameRateController.cpp:
(WebCore::CCFrameRateController::setActive):
(WebCore::CCFrameRateController::onTimerTick):
(WebCore::CCFrameRateController::didBeginFrame):

Identifier: 112190@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125939 268f45cc-cd09-0410-ab3c-d52691b4dbfc