Intermittenly, many WebKit2 tests have results from the wrong test compared to the test just run, giving false failures. https://bugs.webkit.org/show_bug.cgi?id=94505
Reviewed by Ojan Vafai.

It looks like if the webprocess times out, WTR may tell NRWT
that the test completed, but not properly reset its internal
state, and then return stale output for subsequent tests.

This patch modifies NRWT temporarily to check for
"Timed out waiting for final message from web process" in stdout
and treat that as a timeout (and thus kill WTR); this seems
to solve the cascade of failures, but of course there's probably
still a bug in WTR that needs to be fixed.

* Scripts/webkitpy/layout_tests/port/driver.py:
(Driver.run_test):

Identifier: 112302@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@126062 268f45cc-cd09-0410-ab3c-d52691b4dbfc