[Chromium] REGRESSION(r126534) time-multiple-fields-appearance-pseudo-elements.html failing on Win7 https://bugs.webkit.org/show_bug.cgi?id=94898
This patch changes parameters to test expectation generator to make
actual result and test expectation to be matched.

Root cause of this regression is the test expectation generator for
Chromium-Win generates second field with value "56" instead of "00".

On Chromium-Linux and Chromium-Mac, the test expectation generator
doesn't generate second field.

* fast/forms/time-multiple-fields/time-multiple-fields-appearance-pseudo-elements-expected.html: Changes parameters to have hour and minute only.

Identifier: 112764@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@126545 268f45cc-cd09-0410-ab3c-d52691b4dbfc